#22: GET swot filter by membership
