+++ swot_members/fixtures/members.json
@@ -38,5 +38,45 @@
       "member_id": 5,
       "swot_id": 8
     }
+  },
+  {
+    "model": "swot_members.SwotMember",
+    "pk": 6,
+    "fields": {
+      "created": "2018-09-04",
+      "added_by_id": 5,
+      "member_id": 5,
+      "swot_id": 2
+    }
+  },
+  {
+    "model": "swot_members.SwotMember",
+    "pk": 7,
+    "fields": {
+      "created": "2018-09-04",
+      "added_by_id": 5,
+      "member_id": 5,
+      "swot_id": 3
+    }
+  },
+  {
+    "model": "swot_members.SwotMember",
+    "pk": 8,
+    "fields": {
+      "created": "2018-09-04",
+      "added_by_id": 2,
+      "member_id": 2,
+      "swot_id": 4
+    }
+  },
+  {
+    "model": "swot_members.SwotMember",
+    "pk": 9,
+    "fields": {
+      "created": "2018-09-04",
+      "added_by_id": 5,
+      "member_id": 5,
+      "swot_id": 5
+    }
   }
 ]
